- floor autofills when adding a node
