+++ apps/frontend/src/components/EditNodeForm.tsx
@@ -78,7 +78,7 @@ export default function EditNodeForm(props: editNodeProps) {
           {
             node_id: nodeInfo.ID,
             longName: nodeInfo.longName,
-            floor: nodeInfo.floor,
+            floor: props.node.getFloor(),
             nodeType: nodeInfo.nodeType,
             shortName: nodeInfo.shortName,
             x_c: props.node.getX().toString(),
@@ -194,13 +194,6 @@ export default function EditNodeForm(props: editNodeProps) {
           onChange={handleNodeEdits}
           variant="filled"
           label="Long Name"
-        />
-        <TextField
-          value={nodeInfo.floor}
-          name="floor"
-          onChange={handleNodeEdits}
-          variant="filled"
-          label="Floor"
         />
         <TextField
           value={nodeInfo.nodeType}
