+++ apps/frontend/src/objects/getFloorByImage.ts
@@ -1,0 +1,25 @@
+export function getFloorByImage(imgName: string): string {
+  // split the image path name by /
+  const parts = imgName.split("/");
+  // obtain the last part (e.g., 00_thelowerlevel1.png)
+  const lastpart = parts[parts.length - 1];
+  // split based on underscore
+  const splits = lastpart.split("_");
+  //take the number
+  const numbers = splits[0][1];
+
+  if (numbers == "0") {
+    // get only the name without the .png
+    const onlyName = splits[1].slice(0, -4);
+    if (
+      onlyName.includes("thegroundfloor") ||
+      onlyName.includes("thelowerlevel1")
+    ) {
+      return "L1";
+    } else {
+      return "L2";
+    }
+  } else {
+    return numbers as unknown as string;
+  }
+}
